some refactoring

+++ FinanceTracker.Data/Entities/Transaction.cs
@@ -11,7 +11,7 @@ namespace FinanceTracker.Data.Entities
         [Key]
         public int Id { get; set; }
         public int CategoryId { get; set; }
-        public int Amount { get; set; } = 0;
+        public decimal Amount { get; set; } = 0;
         [StringLength(100)]
         public DateTime Date { get; set; }
         public int UserId { get; set; }
